refactor: detached action file names
