+++ src/binary-window/glass/index.js
@@ -1,5 +1,5 @@
-import { default as actionModule } from './action';
-import { default as dragModule } from './drag';
+import actionModule from './action';
+import dragModule from './drag';
 import closeAction from './action.close';
 import minimizeAction from './action.minimize';
 import maximizeAction from './action.maximize';
